fix: address reviews
